Safari-611.1.10/LayoutTests/imported/w3c/IndexedDB-private-browsing/idbindex_openKeyCursor2-expected.txt


PASS IDBIndex.openKeyCursor() - throw InvalidStateError when the index is deleted
